Disable hg types derivation for git repo in unit tests
Summary:
## This stack
Continuing work to support backsyncing submodule extensions.

Most of it will be improving and adding tests to cover the edge cases.

## This diff
The small repo was imported from a git repo and has submodule file changes, so we need to disable hg types derivation (and filenodes bc it depends on hg).

Reviewed By: RajivTS

Differential Revision: D61015476

fbshipit-source-id: 07428088c2d660316ed28f6c5ac5635b1a3ff0e0