Disable FileNodes backfill check for repos that don't derive it
Summary:
## This stack
Continuing work to support backsyncing submodule extensions.

Most of it will be improving and adding tests to cover the edge cases.

## This diff
Skip a filenodes backfill check during pushrebase if filenodes derivation is disabled.

**More context**
When updating unit test to cover backsyncing of submodule changes, I had to disable derivation of hg types in the repos imported from git, because they have submodule changes. Because filenodes depend on hg types, we also have to disable filenodes.

After doing that, pushrebase operations in these repos failed because it tried to derive FileNodes.
I found that this came from this backfill check, which is not needed for the repos that don't have FileNodes enabled.

Reviewed By: RajivTS

Differential Revision: D61116447

fbshipit-source-id: 64ee1ea7d9e828030a1dafacebf5f71136a06fe8