Functional approach to create ExpectedChangeset in unit tests
Summary:
## This stack
Continuing work to support backsyncing submodule extensions.

Most of it will be improving and adding tests to cover the edge cases.

## This diff
Introduces a better way to create instances of `ExpectedChangeset` in unit tests.
Instead of having one function that you pass Regular and Deletion file changes in order, use a functional approach, where we can call `with_regular_changes` and/or `with_deletion_changes` whenever each one is needed.

This is better because:
- No need to pass an empty vec for types of changes that are not needed.
- More readable which kind of changes are being added (instead of relying on comments).
- **More extensible**. In the next diff I'll start adding file changes of type `GitSubmodule`.
  - I noticed that I would need to either add another method to created instances with them, which would add a lot of duplication, or update `new_by_file_change` to take these chages, which would need to update all the existing callsites to pass empty vectors.

Reviewed By: RajivTS

Differential Revision: D61121287

fbshipit-source-id: 99df617253e1c10331f58035dd060d0ef44683aa