progress: report when an inode is finally updated
Summary:
As a first step for reporting progress, this diff increases a counter every time an inode update is complete. Later the following be added:
- Getting an inode count total, as an objective to be completed
- Reporting progress outside of inodes
- Counting inodes as completed early via short-circuiting / other skipping means

The objective of this diff is to make Sapling able to do something with the progress reported by Eden, which is tested in D60385451. Reporting the current checkout mode by Eden is not deemed as necessary since Sapling should know which mode it asked EdenFS to perform checkout with.

Continuously calling this thrift entry point will be gated through Sapling's config

Reviewed By: kmancini

Differential Revision: D60385447

fbshipit-source-id: b0d6f5170d5a1508764c53470fa38f4ccf409657