mercurial/mutation: switch to inner join for selecting predecessors
Summary: Currently we use left join for selecting the predecessors of a commit.  This should ordinarily work, as it should be a constraint that a commit always has at least one predecessor if it has an entry.   However, in the instance where the predecessor rows are missing, left join returns NULL, which is an error.  It's better to ignore this incomplete entry entirely.  We can do that by switching to an inner join.

Reviewed By: YousefSalama

Differential Revision: D61125961

fbshipit-source-id: 3f9223288f9ca3eb79a0cbe40d25934ec38c2521