drop the LFS information when rewriting the commit
Summary:
Right now commit rewriting doesn't copy non_canonical LFS pointer between
repos, if we were copying the LFS data as-is we might get into situation where
we lack the pointer data. We can change this in the future but dropping the LFS
info feels like something that's always safe to do until we are able to make
sure we're also copying non-canonical pointer contents between repos etc..

I'm also not sure what do we want to do with LFS when rewriting commits: should
we leave them as-they-are? Should we use some setting from target repo to make
the LFS decision again? Let's not make that decision now.

Reviewed By: gustavoavena

Differential Revision: D61021070

fbshipit-source-id: 67ae23bb4216c7c57e80d32b8a43b90cc0af14b4