initialize repo configs on startup
Summary:
I don't think we have a good reason to avoid loading configs. We already depend
on configerator in remote_gitimport (for justknobs) and if we keep the amount
of things we initilize reasonable we might be able to use the familiar repo
object for handling things like repo identity and repo configs.

Reviewed By: markbt

Differential Revision: D60951033

fbshipit-source-id: deefb1ec608f40e29ee119bd9917b609707980e3