Unit tests for submodule expansion backsyncing
Summary:
## This stack
Continuing work to support backsyncing submodule extensions.

Most of it will be improving and adding tests to cover the edge cases.

## This diff
Now that we have **some support** for backsyncing submodule changes, let's test this code in the existing unit tests.

Currently, all backsync unit tests directly ran submodule expansion validation to test that logic.

Now, we'll update all tests to backsync the changeset, which will **always run validation**.

This diff updates the only exiting unit test that's supposed to pass validation and successfully backsync the changeset.
To run backsyncing in this test, we properly mock the JK and introduce a helper to create the test data to reverse the commit syncer.

In the next diff I'll update the other unit tests, which will also use this helper.

Reviewed By: mitrandir77

Differential Revision: D61004023

fbshipit-source-id: 8fb886bf3293459e1314b8da621cae1b1d383d26