hg/cas: make hg binary more flexible to include features like cas
Summary:
So while we experimented with the cas client we had it as an isolated target behind a buckconfig `is_experiemntal_cas_build`, this allowed for a risk free development since breaking it wouldn't affect any existing ops or checks. Since this is more mature now we should add it to the build graph.

There's many ways to do this, I chose the less complicated one that resembles the initial approach by muirdm so we don't make this more complex than it is. It basically makes the hg binary generic while keeping the internal version as the "main" version but allows mirrors to it with separate features. Either cas or whatever we want to add later.

The main benefit of this is that still allows for a separation of the hg binary while adding coverage for `hg_cas` on citadel. Eventually we want to make everything that runs with `hg` run with `hg_cas` before we make the switch.

The bad side is that this will make `hg` builds slow for a couple of days while we roll it out since well, it invalidates the whole cache, but I guess that's okay given we're rolling out a "new" hg.

Reviewed By: sggutier

Differential Revision: D61126754

fbshipit-source-id: 33d4f06b1f5001086dbd1a7c0b4e479e040b6ca0