add clear API for ScmStatusCache
Summary:
# Background

We have a plan to optimize ScmStatusCache by ignoring the file changes under .hg folder when counting the Journal sequence change.

# Tech Detail

In this diff, we added an clear() API for `ScmStatusCache` that can clear both the promise map as well as the internal `ObjectCache`. This will be triggered when we detect an update of the working copy.

Usually this is not an issue since when updating the root hash, `Journal` will be notified and the sequence ID will be adcanced. But this is necessary if we decide to reuse the caced results under some special cases: e.g. there is only `.hg` changes from a range of journal delta.

This change guard us from reusing something that's not supposed to be reused.

Reviewed By: kmancini

Differential Revision: D60862425

fbshipit-source-id: 9f113f19617ceb72ad02cb696e9badd988a42e98