restore the behavioro of a test helper function
Summary:
# Backgroun

Changing this back to 0.0 to restore the original behavior of the existing tests. I added this extra parameter in one of my previous diff but I realized that setting it default to 1.0 is causing the existing tests to do the wait-in-loop assertion instead of a one-time assertion.

Thus changing it back so only my new tests will be impacted.

Reviewed By: genevievehelsel

Differential Revision: D60865716

fbshipit-source-id: ce7cafed7250d4d26db354dcfa0adbdec92c6ad1