smarter get of ScmStatusCache by ignoring .hg folder changes
Summary:
# Background

We have a plan to optimize ScmStatusCache by ignoring the file changes under .hg folder when counting the Journal sequence change.

# Tech Detail

In this diff, we are making the `get` method of `ScmStatusCache` smarter by utilizing `isSequenceValid` helper function when we see a mismatch between the cached sequence and the current sequence. This optimzation applies for both the promise map and the internal `ObjectCache`.

# Other Changes

- unified the API fot `get` and `insert` by explicitly requiring a sequence ID to be provided.
- added a logger for `ScmStatusCache` and added debug logging msg  for core ops

Reviewed By: kmancini

Differential Revision: D60869910

fbshipit-source-id: a8afd834ffa9eb2ae9ad9f501887ec62da2a120c