Read traffic routing to Metagit if repo is non-Mononoke
Summary: Sometime down the line, we would want to have the `route-to-mononoke` JK enabled by default instead of disabled by default. This diff relies on the ShardManager status of a repo to determine if the traffic should be directed to Mononoke or Metagit, i.e. if the repo is not present as a sharded repo in Mononoke then it has to be redirected to Metagit. Note that the current diff only applies for read Git proto traffic. A follow-up diff will cover read Git replication traffic and later in the stack I will cover write Git proto traffic.

Reviewed By: gustavoavena

Differential Revision: D61123984

fbshipit-source-id: ea755992a16b57241e2b24979f88afca2649fc28