commit cloud: add owner identity to commit cloud context
Summary:
Inferring a workspace owner is slow since we need to query the employee service.

In `share_workspace`  we have  to check if the user is allowed + create the acl (which also requires the owner identity), which would make us do this query twice, far from ideal. So I'm storing it at the commit cloud ctx so it can be reused later

Reviewed By: RajivTS

Differential Revision: D61139054

fbshipit-source-id: a372af77f7c4261006c208e2a742d33f27eb4e07