mononoke_api primitive to update submodule expansions in large repo
Summary:
## This stack
With the support for backsyncing submodule expansions changes in the large repo, we need to provide to users a way to generate valid submodule expansion changes in the large repo.

This stack introduces the SCS method and command that will let them do so.

The command will be something like:

```
scsc update-submodule -R <LARGE_REPO> --base-commit <LARGE_REPO_COMMIT> --submodule-expansion-path <PATH> --new-submodule-commit <SUBMODULE_COMMIT>
```

## This diff
Mononoke API method that has the core logic to generate the commit updating the submodule expansion.

This was not very complicated because we use existing primitives.
The logic is basically:

- Take a submodule path and a new submodule update, which can be to delete the submodule (i.e. its expansion) or point it to a new commit.
- Find the small repo being affected based on the path of the submodule.
- Create a commit in the small repo, with the GitSubmodule file change.
- Sync it to the large repo and return that commit.

Reviewed By: andreacampi, mitrandir77

Differential Revision: D60968869

fbshipit-source-id: 061cbf0cd46bee1ad7493d5d10502aea87c94645