Implement SCS client command to update submodule expansion in large repo
Summary:
## This stack
With the support for backsyncing submodule expansions changes in the large repo, we need to provide to users a way to generate valid submodule expansion changes in the large repo.

This stack introduces the SCS method and command that will let them do so.

The command will be something like:

```
scsc update-submodule -R <LARGE_REPO> --base-commit <LARGE_REPO_COMMIT> --submodule-expansion-path <PATH> --new-submodule-commit <SUBMODULE_COMMIT>
```

## This diff
Implements the SCS client command. I actually implemented this after the server method (D60764427), but decided to publish it before since it's simpler and the client will be used in the integration test in D60764427.

Reviewed By: andreacampi, mitrandir77

Differential Revision: D60827576

fbshipit-source-id: 46d1a69c1fe58f86a59784ca58be051649f32fa8