Move helper to util module
Summary:
## This stack
With the support for backsyncing submodule expansions changes in the large repo, we need to provide to users a way to generate valid submodule expansion changes in the large repo.

This stack introduces the SCS method and command that will let them do so.

The command will be something like:

```
scsc update-submodule -R <LARGE_REPO> --base-commit <LARGE_REPO_COMMIT> --submodule-expansion-path <PATH> --new-submodule-commit <SUBMODULE_COMMIT>
```

## This diff
Extract helper that will be reused in D60827576

Reviewed By: mitrandir77

Differential Revision: D61125728

fbshipit-source-id: edc780721d703e08e37191bffb278d960ada31ed