Fix focus not regaining when tabbing back and forth
Summary:
Previously, if you are writing a commit message, then cmd-tab away from vscode and cmd-tab back, it would have lost focus on the commit info view field. In fact, it will have lost focus on ISL entirely, requiring a mouse click to regain.

This appears to be a vscode webview bug, as this doesn't repro in the browser. VS Code webviews are weird about focus in a few situations.

We can write our own custom mitigation for this. The idea is to save the currently focused element when the window is blurred. Then, when the window regains focus, we refocus our saved element. This technique works for any focus-able element, not just textareas.

This seems to work well in my testing, though this is one of those features that is a bit hacky, by saving elements and focusing them without any context about what that thing was.

Reviewed By: alichtman

Differential Revision: D61145666

fbshipit-source-id: 2ac9dd3212200d0beea3c0fac5e21951c365c868