fix a bug where we could not distinguish between a parse error and "--help"
Summary:
we are expecting "eden help" and "eden --help" to work and we are expecting the return code to be 0.

However, if we try "eden --help" today it will return code 64 although the output seems correct (if we are ok to see python output for one and rust output for the other).

The reason we are seeing a return code 64 is that when we fallback to python, the python process is actually returning 64 as an exit code if "--help" is provided. [code pointer](https://www.internalfb.com/code/fbsource/[5188e3d34ef2]/fbcode/eden/fs/cli/main.py?lines=2795)

The reason we added it was because `parse_args` could fail with parse error and the exception was a SystemExit(2). But that `except` can also capture `SystemExit(0)` when the subcommand completes directly when `--help` is requested.

So we should distinguish between the exit code here.

Reviewed By: MichaelCuevas

Differential Revision: D61054047

fbshipit-source-id: 75b9a5b56ed28899278fee1a41185ef301f3c320