progress: add an external progress bar for EdenFS gotos
Summary: EdenFS goto has multiple phases, both within EdenFS and within Sapling. In order to get users an idea of how much things are left to do let's introduce an external progress bar for the phases in a checkout. At the moment this will have at most one or two phases with a normal/forced goto, but in the near future will contain more

Reviewed By: muirdm

Differential Revision: D61000484

fbshipit-source-id: 5c0f2005426270efe01902082ec37bfbd48b5649