Extract logic that handles updates to submodule expansion (i.e. no deletions)
Summary:
## This stack
Continuing work to support backsyncing submodule extensions.

Most of it will be improving and adding tests to cover the edge cases.

## This diff
Extracts the existing logic, which handles **updates** to the submodule expansion (i.e. updating the submodule pointer), into a separate function.

In the next diffs we'll introduce support for deletion.

Reviewed By: mitrandir77

Differential Revision: D61200897

fbshipit-source-id: afa2eb16e6428b1497907ec8f7f9bdd2340a7347