Support for backsyncing submodule expansion deletions
Summary:
## This stack
Continuing work to support backsyncing submodule extensions.

Most of it will be improving and adding tests to cover the edge cases.

## This diff
Adds support for backsyncing submodule expansion deletions.

Similarly to compacting regular changes, this isn't very complex because we always run submodule expansion validation before compacting, so we can make some assumptions.

**However**, when it comes to deleting the submodule metadata file, there is one important case to cover.
It's described in T187241943 and in the doc comments of this diff.

The **TL;DR** is: deleting just the submodule metadata file is a valid change that could be backynced to deleting the submodule in the small repo. **But**, this would also add the entire submodule expansion to the small repo working copy.

I believe that it's unlikely that users will do this **intentionally**. Even if they did, this would result in a huge bonsai, which is not ideal.

So let's not let users do that and only backsync a bonsai that's deleting the metadata file and the full expansion, **which is what the SCS and Sapling commands to update submodules will do**.

Reviewed By: mitrandir77

Differential Revision: D61201565

fbshipit-source-id: c9a7d47a6b73dea15942ee6e19b611289a6c5c54