Honor SoT in Location Service when routing git push requests
Summary: Currently, even though Mononoke might be the Source of Truth for some Git repo, any pushes to such a repo do not get auto redirected to Mononoke. To do so requires the use of a custom header. This diff makes the necessary changes so that the location service relies on the Git Source of Truth table to decide if the request should be routed to Mononoke or Metagit. This is necessary for seamless migration of Git traffic from Metagit to Mononoke.

Reviewed By: singhsrb

Differential Revision: D61264245

fbshipit-source-id: 5eb93d9cc84273f28f485f993176fef06982eb42