Provide link to hooks wiki in case of Git hooks failure
Summary: Currently, we just have the error message whenever a hook fails. As good UX, it might be useful to redirect the user to the hooks wiki to see how hooks can be bypassed or disabled for future pushes. This should also help reduce the user posts that the oncall would need to deal with.

Reviewed By: mzr

Differential Revision: D61267512

fbshipit-source-id: c6d948fbc8039e2328dd92cf8371f2de95f15f65