Provision for accessing git commits ids from bonsais in hooks
Summary: As part of hook migration from Git to Mononoke, we want to migrate hooks that require access to the Git Shas of the commits rather than the bonsai hashes. This diff adds the `get_git_commit` method that can be used for any Git repo to convert a bonsai commit in that repo into a git commit id.

Reviewed By: mzr

Differential Revision: D61268531

fbshipit-source-id: aecbd176fb3216d5a4e84c9347c68249d6faa0ea