More clarity on MissingFilesForInodes
Summary:
# Context
MissingFilesForInodes happens when a file doesn't exist when the inode says it should. It is usually fixed by creating and removing a blank file at the path and having eden pick this up and recreate the file

# This Diff
Recently a check for the files being recreated was added. This check is somewhat poor in the following aspects that this diff aims to fix:
- It runs immediately after perform_fix possibly leading to a race condition on eden recreating the file
- The error messaging is somewhat unclear

# Technical Details
Make it explicit that the error messaging from the check is from the check and not from perform_fix
Add a short sleep after perform_fix to let eden update the filesystem
Inform the user about the time lag and ask them to check if files are still missing

Differential Revision: D60970702

fbshipit-source-id: 4b5b142918572c73907692c81bf2dfee4959efa3