Extract common pattern in unit test to helper function
Summary:
## This stack
Continuing work to support backsyncing submodule extensions.

Most of it will be improving and adding tests to cover the edge cases.

## This diff
Another refactor: I'm trying to simplify the unit tests, finding common patterns and reducing copy-pasting.
This diff extracts another common pattern to a helper function, where we sync a changeset, get all the changesets in the target repo and print them (to debut tests if they fail), then derive all types for these changesets.

Having this in a helper that's reused not only makes the unit tests smaller and easier to read, but also automatically apply any improvement (e.g. debug logs) to all of them.

Reviewed By: lmvasquezg

Differential Revision: D61268348

fbshipit-source-id: b2ea93b38202e97f46003d0468385a0621612e05