Fix checkout config corruption during eden startup doctor
Summary:
# Context
Eden checks for checkout_info when starting up. If the checkout is corrupted, doctor will report this issue. This currently works fine when eden is already up and running, but if it isn't then doctor will fail.

# This Diff
Adds a check to the initial code that gets the CheckoutInfo that will catch the errors thrown by _read_config().

Reviewed By: genevievehelsel

Differential Revision: D60986586

fbshipit-source-id: aa5e120919e56a11d473595828e1b067086c111b