metrics: move blackbox counter logging to Rust
Summary:
This way they get logged for native commands, and we now have all the counter handling consolidated in Rust.

I followed the advice of the TODO in Python and made the blackbox logging not filter based on devel.skip-metrics.

Python was using `pprint.pformat(..., width=1024)`, but I did not reproduce the max width. Didn't seem important for the blackbox.

Reviewed By: quark-zju

Differential Revision: D61346734

fbshipit-source-id: 473c27a4f432655d6d2983e5d69d9fcbb58456f0