repo: add TreeStore wrapper with cache
Summary:
Add a TreeStore impl that wraps another TreeStore, adding an in-memory cache.

This caching targets manifest-tree activity, which is just the KeyStore part of things (i.e. fetching raw hg manifest content). It could be extended to support AUX data, but that would be more complex and is probably not needed, at least for Sapling.

Reviewed By: zzl0

Differential Revision: D61291343

fbshipit-source-id: 702fe22ea5e9233735b3cdcddc51506fec64cf82