repo: add caching to tree resolver
Summary:
The TreeResolver is how code like "status" and "checkout" initially fetch TreeManifest objects. Previously, there was no caching, so if the same manifest was fetched twice, each subsequent tree access as the manifest was traversed would go to the underlying TreeStore. This may have been fine for trees in the hgcache since the data would end up mmaped into memory, but is less fine for CAS (and maybe Git?).

Python has tree caching in the treemanifest extension at the manifestctx level. The caching I've added here operates at the tree level, so will be effective across different manifests that contain the same trees.

I chose to only cache for the TreeResolver, basically, because I did not want the caching to take effect for the TreeStore used by EdenFS. EdenFS has its own in-memory caching, so I didn't want to add another layer.

Reviewed By: zzl0

Differential Revision: D61291344

fbshipit-source-id: 45607953b9f66da7789502e37f81243d24b6c40c