commit cloud: add rename_workspace queries
Summary: When we rename a workspace we need to update _all_  of the tables with the new name, so adding the query in all handlers, next diff will use it. The only weird thing here is bookmarks which is named `bookmarks` in prod but `localbookmarks` in tests to not conflict with the actual mononoke `bookmarks` tables.

Reviewed By: gustavoavena

Differential Revision: D61333109

fbshipit-source-id: 1af085f951c191770c7b9dbf2eb4f2313d007524