store: kill fncache
Summary:
This was a cache for mapping revlog file names to their escaped/encoded versions actually used on disk. We don't support actual revlogs anymore, so we don't need this.

I got rid of most of the path encoding/decoding logic (and tests), except for encodefilename which is used by debugexportrevlog.

The main thing I wanted to get rid of was the "fncache" transaction file (trying to limit stuff done by transactions).

Reviewed By: quark-zju

Differential Revision: D60968569

fbshipit-source-id: 6b25da69a766875d9987d984982785d00439e60c