Provision for immediate read-after-write consistency in Git push
Summary:
There are various scenarios in our CI infrastructure where an immediate read-after-write consistency is expected after a Git push (i.e. if I pushed master from A -> B, then pulling immediately after the push should show master at B or later). This is not the default behavior for Mononoke Git server since we rely on WBC to reduce the load on bookmarks DB.

In situations where this is needed, we can provide a best-effort mechanism to wait for the push to get updated in the WBC so a follow up request to the same server would have immediate read-after-write semantics

Reviewed By: andreacampi

Differential Revision: D61471388

fbshipit-source-id: 1fddd2a44b7e464224eb1d4241189f6ce98cbe6a