testing: fix _get_record_metalog has no attribute 'cache_clear' error
Summary:
The _get_record_metalog() function might be using `functools.lru_cache` before,
so it calls cache_clear() on the function object to clear the cache. But now
the _get_record_metalog() does not use it, and will cause AttributeError when
running 'sl .t --record' twice.

The `sl .t --record` is not a long running process, so it is probably okay just
remove the line calling cache_clear()

Reviewed By: muirdm

Differential Revision: D61476399

fbshipit-source-id: e31e7c85553b861fb5fdd993c8688a8a160d5e13