dag: skip negative cache optimization when adding virtual heads
Summary:
The negative cache optimization is to avoid sequential remote lookup, like when
inserting B001..B999:

    B999
    :
    B001
   /
  A  <- Lasy (master) group head.
  :
  :  <- Lazy vertexes here might be named "B001", not confirmed yet.

Instead of following the heads from B999 and asking the server lots of questions like
"Is B999 an ancestor of A?", "Is B998 an ancestor of A?", ... The optimization
asks the server about B001 and expect the server to answer "not an ancestor of A"
then add B001 to B999 to the "negative cache" to avoid O(N) remote lookups.

This optimization is not necessary for the virtual group and causes some noisy
tracing logs. Let's disable it for the virtual group.

Reviewed By: zzl0

Differential Revision: D61419239

fbshipit-source-id: e077e9053901c7d31836f4b001ff44b2db9ac857