git_server_push_poller: update fingerprint state in database correctly
Summary:
We noticed that the draining a host with `mzr-test` as leader was not making
progress complaining about the push epoch being set.

This is because the push epoch is updated by the metagit push tailer to
indicate new pushes were made to the repository.

Let's update the push poller to reset the fingerprint state to unblock the
draining.

Reviewed By: YousefSalama

Differential Revision: D61533950

fbshipit-source-id: e4fede0dc1dc4a0f70ade3d5a70c8164f20b8a24