mononoke_x_repo_sync_job: set ClientInfo correctly in CoreContext
Summary: Currently in `mononoke_x_repo_sync_job` when constructing `CoreContext`, we build `ClientInfo` correctly with the appropriate `ClientEntryPoint`, but we use it only for mutating scuba and we don't store it inside of `CoreContext`. This means that SQL queries don't get annotated with client info. This diff fixes this.

Reviewed By: RajivTS

Differential Revision: D61537421

fbshipit-source-id: f64bc156303402b4710cc61f97026415934f46ff