import pySocks to allow for fwdproxy usage later
Summary:
All connections between prod and the internet are routed through fwdproxy (https://www.internalfb.com/intern/wiki/Traffic/Proxygen_Services/ForwardProxy/). The Mononoke on AWS will recieve pushes from the hg sync running in prod.

For that reason, hg must have an option to use this fwdproxy. Fwdproxy uses IPv6 so auth_proxy option did not work for this case.

PySocks is a Python libary that allows to set up an http proxy, we ar eimporting it before using it in the next diff

Reviewed By: mitrandir77

Differential Revision: D61470104

fbshipit-source-id: 7438b4325d7deb8f310b01ef96d5fca13b352bd7