Fix for insidious packfile parsing error during Git push
Summary:
Ever since we rolled out Git write protocol for `mzr-test` repo, every few weeks we would see a server 500 response when trying a simple git push. The push would appear to be completely ordinary yet the problem would occur intermittently and then completely disappear.

There was a long and winded investigation which eventually led to the root cause (I will make a post for it). The reason for the intermittent-ness of the problem was that the error depended on the structure of objects in the packfile which in turn depended on the sequence of refs which the server advertises before the push. So far, that was random which led to the on-off-on behavior of this issue.

The root cause of the error was a single line of code that was trimming whitespaces from the request payload. This was a valid strategy for Git read path because Git introduces unnecessary spaces in some cases but is not required at all in Git write path. Removing the offending line fixed the issue which can be observed through the integration test. Additionally, I have now imposed an ordering on the advertised refs so that we have predictable behavior for future errors.

Reviewed By: andreacampi

Differential Revision: D61538850

fbshipit-source-id: fb7be903a7ef51802fc379455a32949857521547