Remove "Beta" from chunk selection
Summary: This has been in "Beta" for too long. We haven't made changes to it, but people will lose trust in our labeling if something is forever beta. Let's just call it non-beta now, even though it has some rough edges. We want to get feedback on it anyway, so I don't want people to think they can't submit bugs about it.

Reviewed By: quark-zju

Differential Revision: D61488295

fbshipit-source-id: a18bea1907146a85690359cdd91d3cd7ba4ca6ef