edenfs_config_manager: add ability to read test configs from Manifold
Summary:
This diff adds a third testing option to `eden_config_manager`: `--manifold-key`. When `--manifold-key` is set, the config manager will try to find a materalized_config in the `edenfs_test_config` Manifold bucket and use that for config generation. This is helpful for testing on OD, but could be useful too for testing on machines without a Configerator repo locally.

Right now, only source_control team members can upload materalized_configs to that Manifold table, but this new config testing method should make it easier for partner teams to test their changes as well by requesting temporary ACL access to the bucket.

This is mirrored in Sandcastle in D61002494

Reviewed By: kmancini

Differential Revision: D61072490

fbshipit-source-id: dd1658cd24893fc707228de4e04171957de2d392