dag: disable hex prefix match for virtual vertexes
Summary:
The virtual vertexes usually use faked names. Right now they are just wdir and
null. In the future we might want to generate a bunch of other vertexes. Those
generated names should be considered as implementation details and not exposed
in prefix match. This also matches the Python-side prefix matching expectation.

Reviewed By: muirdm

Differential Revision: D61497945

fbshipit-source-id: 76b825d33d7f49983364297a7be20ba17fe95d0e