Align rightmost wrapped items to the right in the top bar
Summary:
The top bar has tools like the pull button, bug button, settings, and refresh. These are broken into left and right tools on large screens.

On small screens, this wraps the right group onto a new line to fit without scrolling. However, it was wrapping to the left side, due to how flexbox works by default.

We can work around this by adding a flex spacer on the left, and `flex-grow`-ing the entire group.

Reviewed By: sggutier

Differential Revision: D61568210

fbshipit-source-id: 94bcddea6ad7fc66003d97e260dc3a4bca1c854e