scmstore: avoid duplicate aux cache writes in treestore
Summary:
The two "remote" sources of trees are SaplingAPI and CAS. SLAPI writes tree content back to the hgcache, where CAS does not. This means we will hit CAS every time we fetch a tree, so it doesn't make sense to write the aux data from the tree to the aux cache every time (since it will result in a lot of duplicate writes).

Avoid duplicate writes by checking if the aux caches already contains the aux data. This isn't great, but we don't have a good way to know that whether we should write the aux data. An alternative would be to store an extra bit of knowledge in the parent tree's aux data indicating whether we have written its children aux data, but that runs the risk of caches rolling over and us not writing file aux data when we should.

Reviewed By: quark-zju

Differential Revision: D61494230

fbshipit-source-id: 4d6c000117aae787a51a226cecfaa5d3e7a9404e