pyrevisionstore: use the shared caching tree store
Summary:
Embed the caching tree store in the treescmstore object so Python can benefit from it.

In particular:
- Make the "get()" and "prefetch()" methods use the caching tree store, if present. This handles adhoc usage from Python.
- Make the treescmstore to_dyn_treestore impl prefer the caching store, if present. This causes treemanifest objects constructed in Python to use the caching store instead of the normal store.

Compared to the existing Python treemanifest caching:
- The caching tree store caches at the tree-level instead of the manifest level, so it can be shared across different manifests that have trees in common.
- The caching tree store is (now) shared across Rust and Python.

This is sort of an awkward place to put the cache, but the Python doesn't currently use the TreeStore trait object directly, so sticking it inside the treescmstore is the most useful place. Note, however, that this means the cache won't work for git or eager repos in Python land (since those storage formats don't use treescmstore).

Reviewed By: quark-zju

Differential Revision: D61494229

fbshipit-source-id: bdc1685397386a243e26c682d9b8412425ae187c