Implement retry for GitLocator service in case of transient issues
Summary:
Over time, we have observed that whenever the Mononoke Git server tasks restart as part of conveyor push, the requests at that time fail with the following error:

```
LocationHandler error: HttpError { error: Host twshared47558.05.cln5.facebook.com present in the Shard Manager Mononoke Git tier but not in regular Mononoke Git tier, status_code: 500 }
```

This error occurs when we get a host from ShardManager that is not present in the Mononoke Git SMC tier. This is a timing problem and is very short lived. It also only happens during [TW restarts or conveyor pushes](https://fburl.com/scuba/slb_500/h7ahs500).

This diff adds conditional retry for the GitLocator crate so if the request is made during conveyor rollout or TW restart, we can wait a while to select the correct host for the required repo.

Reviewed By: andreacampi

Differential Revision: D61549300

fbshipit-source-id: b2d9ec48163921d90dae859f66404423bfe0bdac