Refactor the Git protocol crate
Summary: The git protocol crate contains `generator.rs` file which is humongous at this point and heavily against SRP. This diff tries to refactor the git protocol crate so its a bit easier to read and functions are categorized based on their intent. Futher refactorings are possible but this is good enough for now.

Reviewed By: YousefSalama

Differential Revision: D61595145

fbshipit-source-id: 805e08011026856f5f3bcbfe70af553c8694620d