Integration test demonstrating failure in handling submodule in root commit
Summary: When importing a Git repo in Mononoke, we derive GitDeltaManifest for it as well. `GitDeltaManifest` for a commit should include entries for blobs and trees but not for submodule entries which show up as commits. However, in the current implementation when deriving GitDeltaManifest for the root commit, we add an entry for every item in the corresponding Git Tree without even checking if the item is a submodule or not. As a result, we end up with an erroneous GitDeltaManifest that leads to failing clone as demonstrated by this integration test

Differential Revision: D61601189

fbshipit-source-id: 15187a0bcc8053f2cc774752f92213ee95fe0244