build: add corepack support (#935)
Summary:
[Corepack](https://nodejs.org/api/corepack.html) is an experimental package
manager manager for Node.js. It is a part of the Node.js project and is
intended to be a more secure and reliable alternative to other package manager
managers. This diff adds support for Corepack to the project.

Pull Request resolved: https://github.com/facebook/sapling/pull/935

Test Plan:
`corepack` comes with every Node.js installation (besides homebrew), so there
are no extra installation steps needed.
-1. If you installed Node.js with homebrew, run `brew install corepack`
0. Because corepack is experimental, run `corepack enable`.  Once it is
stabilized this step will be uneccesary
1. `cd addons/`
2. `yarn`

Now, depedencies will always be installed with the correct version of `yarn`.
In my case, `yarn` defaults to `4.3.1`, but when I run `yarn --version` in the
`addons/` directory, it returns `1.22.22`.

Reviewed By: evangrayk

Differential Revision: D61603154

fbshipit-source-id: 0ea74ebcf05caf358731f31757ce835919ff9562