Add button to mount additional workspace folders in VS Code
Summary:
It was a little unclear how ISL cwd options relate to repositories you have on your machine. If we add a button to mount additional workspace folders, it becomes much more clear that it's vscode workspace folder-based. We can further explain in the tooltip.

This only applies on vscode, though this would be a nice UX for other platforms to add other cwd options too.

Reviewed By: muirdm, alichtman

Differential Revision: D61606753

fbshipit-source-id: 338ce9ec1319ff74a437b4c1b38dd7b8704503cd