cmdutil: refactor registerdiffgrafts() function
Summary:
The motivation is that we want to do paths check when graft created no changes to commit. One of the common reasons of no changes to commit is users might provide wrong `from-path`, we'd like to check the path existence when no changes to commit (see D61598640)

This diff refactors registerdiffgrafts() function so it takes from/to paths arguments instead of `opts`. And we will do some validations on from paths outside of registerdiffgrafts()

Reviewed By: muirdm

Differential Revision: D61598643

fbshipit-source-id: 860dd136521ec5f620c122c7b96c676cfd87871c